# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Plateforme Logistique Dakar API
  description: |
    API pour la gestion de livraisons pour petits vendeurs Instagram/Facebook Ã  Dakar.
    
    **FonctionnalitÃ©s principales :**
    - Gestion des livraisons (crÃ©ation, suivi, statuts)
    - Ramassages groupÃ©s par zones
    - Notifications WhatsApp automatiques
    - Gestion financiÃ¨re (Cash on Delivery)
    - Tracking public des colis
    
    **Authentification :** JWT Bearer Token
  version: 1.0.0-MVP
  contact:
    name: Plateforme Logistique Dakar
    email: contact@votreplateforme.sn
servers:
  - url: http://localhost:8080/api
    description: Serveur de dÃ©veloppement
  - url: https://api.votreplateforme.sn/api
    description: Serveur de production
...
security:
  - bearerAuth: []
paths:
...
  /closeur/commandes:
    get:
      tags:
        - Closeur
      operationId: closeurCommandesGet
      summary: Commandes à traiter par le closeur
      description: Liste les commandes Nouvelle / A appeler / Confirmée (file du closeur)
      parameters:
        - name: statut
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/StatutLivraison'
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Liste des commandes du closeur
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommandeCloseur'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
      security:
        - bearerAuth: [ ]
  /closeur/commandes/{id}/appeler:
    post:
      tags:
        - Closeur
      operationId: closeurAppeler
      summary: Prendre la commande en charge pour appel (Nouvelle -> A appeler)
      parameters:
        - $ref: '#/components/parameters/LivraisonIdParam'
      responses:
        '200':
          description: Commande passée à "A appeler"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandeCloseur'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      security:
        - bearerAuth: [ ]
  /closeur/commandes/{id}/confirmer:
    post:
      tags:
        - Closeur
      operationId: closeurConfirmer
      summary: Confirmer la commande après accord du client (-> Confirmée)
      parameters:
        - $ref: '#/components/parameters/LivraisonIdParam'
      responses:
        '200':
          description: Commande confirmée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandeCloseur'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      security:
        - bearerAuth: [ ]
  /closeur/commandes/{id}/prete-a-livrer:
    post:
      tags:
        - Closeur
      operationId: closeurPreteALivrer
      summary: Rendre la commande disponible pour le dispatch (-> Prête à livrer)
      parameters:
        - $ref: '#/components/parameters/LivraisonIdParam'
      responses:
        '200':
          description: Commande prête à livrer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandeCloseur'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      security:
        - bearerAuth: [ ]
  /closeur/commandes/{id}/reporter:
    post:
      tags:
        - Closeur
      operationId: closeurReporter
      summary: Reporter la commande (le client demande un rappel plus tard)
      parameters:
        - $ref: '#/components/parameters/LivraisonIdParam'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentaireRequest'
      responses:
        '200':
          description: Commande reportée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandeCloseur'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      security:
        - bearerAuth: [ ]
  /closeur/commandes/{id}/annuler:
    post:
      tags:
        - Closeur
      operationId: closeurAnnuler
      summary: Annuler la commande (-> Annulée)
      parameters:
        - $ref: '#/components/parameters/LivraisonIdParam'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentaireRequest'
      responses:
        '200':
          description: Commande annulée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandeCloseur'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      security:
        - bearerAuth: [ ]

  # ============================================
  # ADMIN - DISPATCH
  # ============================================
...
  /admin/closeurs:
    get:
      tags:
        - Admin - Equipe
      operationId: listCloseurs
      summary: Lister les closeurs
      responses:
        '200':
          description: Liste des closeurs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MembreResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
      security:
        - bearerAuth: [ ]
    post:
      tags:
        - Admin - Equipe
      operationId: createCloseur
      summary: Créer un compte closeur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMembreRequest'
      responses:
        '201':
          description: Closeur créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembreResponse'
        '400':
          description: Requête invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
      security:
        - bearerAuth: [ ]
  /admin/livreurs:
    get:
      tags:
        - Admin - Equipe
      operationId: listLivreurs
      summary: Lister les livreurs
      responses:
        '200':
          description: Liste des livreurs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LivreurResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
      security:
        - bearerAuth: [ ]
    post:
      tags:
        - Admin - Equipe
      operationId: createLivreur
      summary: Créer un compte livreur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMembreRequest'
      responses:
        '201':
          description: Livreur créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LivreurResponse'
        '400':
          description: Requête invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
      security:
        - bearerAuth: [ ]

  # ============================================
  # LIVREUR - Mes livraisons
  # ============================================
  /livreur/mes-livraisons:
    get:
      tags:
        - Livreur
      operationId: mesLivraisons
      summary: Mes livraisons assignées
      description: Le livreur connecté ne voit que ses propres livraisons (Assignée / En livraison)
      parameters:
        - name: statut
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/StatutLivraison'
      responses:
        '200':
          description: Livraisons du livreur connecté
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommandeLivreur'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
      security:
        - bearerAuth: [ ]
  /livreur/livraisons/{id}/commencer:
    post:
      tags:
        - Livreur
      operationId: livreurCommencer
      summary: Commencer la livraison (-> En livraison)
      parameters:
        - $ref: '#/components/parameters/LivraisonIdParam'
      responses:
        '200':
          description: Livraison démarrée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandeLivreur'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      security:
        - bearerAuth: [ ]
  /livreur/livraisons/{id}/livrer:
    post:
      tags:
        - Livreur
      operationId: livreurLivrer
      summary: Marquer la livraison comme livrée (-> Livrée)
      parameters:
        - $ref: '#/components/parameters/LivraisonIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LivrerRequest'
      responses:
        '200':
          description: Livraison effectuée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandeLivreur'
        '400':
          description: Requête invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      security:
        - bearerAuth: [ ]
  /livreur/livraisons/{id}/echec:
    post:
      tags:
        - Livreur
      operationId: livreurEchec
      summary: Marquer un échec de livraison avec motif (-> Échec)
      parameters:
        - $ref: '#/components/parameters/LivraisonIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EchecRequest'
      responses:
        '200':
          description: Échec enregistré
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandeLivreur'
        '400':
          description: Requête invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      security:
        - bearerAuth: [ ]

# ============================================
# COMPONENTS
# ============================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Token manquant ou invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: AccÃ¨s refusÃ© (permissions insuffisantes)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Ressource non trouvée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  parameters:
    LivraisonIdParam:
      name: id
      in: path
      required: true
      description: Identifiant de la livraison / commande
      schema:
        type: integer
        format: int64
  schemas:
...
    UserRole:
      type: string
      enum:
        - VENDEUR
        - ADMIN
        - CLOSEUR
        - LIVREUR
    StatutLivraison:
      type: string
      enum:
        # Nouveau cycle (Closing + Dispatch)
        - NOUVELLE
        - A_APPELER
        - CONFIRMEE
        - PRETE_A_LIVRER
        - ASSIGNEE
        - EN_LIVRAISON
        - LIVREE
        - ECHEC
        - ANNULEE
        # Ancien cycle (module ramassage - dormant, conservé pour l'historique)
        - EN_ATTENTE_RAMASSAGE
        - RAMASSE
        - EN_ROUTE
        - ECHEC_ABSENT
        - ECHEC_REFUSE
    MotifEchec:
      type: string
      description: Motif obligatoire lorsqu'une livraison passe au statut ECHEC.
      enum:
        - TELEPHONE_INJOIGNABLE
        - CLIENT_ABSENT
        - ADRESSE_INCORRECTE
        - REFUS_CLIENT
        - REPORT_CLIENT
...
    CommandeCloseur:
      type: object
      description: Commande vue par le closeur
      properties:
        id:
          type: integer
          format: int64
        numeroTracking:
          type: string
        statut:
          $ref: '#/components/schemas/StatutLivraison'
        nomClient:
          type: string
        telephoneClient:
          type: string
        adresse:
          type: string
        zone:
          type: string
        produit:
          type: string
        montantCOD:
          type: number
          format: decimal
        dateCreation:
          type: string
          format: date-time
...
    CommandeLivreur:
      type: object
      description: Livraison vue par le livreur (Mes livraisons)
      properties:
        id:
          type: integer
          format: int64
        numeroTracking:
          type: string
        statut:
          $ref: '#/components/schemas/StatutLivraison'
        nomClient:
          type: string
        telephoneClient:
          type: string
        adresse:
          type: string
        pointRepere:
          type: string
        zone:
          type: string
        produit:
          type: string
        montantAEncaisser:
          type: number
          format: decimal
    MembreResponse:
      type: object
      description: Compte d'un membre de l'équipe (closeur ou livreur)
      properties:
        id:
          type: integer
          format: int64
        nom:
          type: string
        prenom:
          type: string
        telephone:
          type: string
        email:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        actif:
          type: boolean
    LivreurResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nom:
          type: string
        prenom:
          type: string
        telephone:
          type: string
        email:
          type: string
        zonePreferee:
          type: string
        actif:
          type: boolean
        nombreLivraisonsEnCours:
          type: integer
          description: Nombre de livraisons Assignée + En livraison en cours
    CreateMembreRequest:
      type: object
      required:
        - nom
        - prenom
        - telephone
        - password
      properties:
        nom:
          type: string
        prenom:
          type: string
        telephone:
          type: string
          example: "771234567"
        email:
          type: string
        password:
          type: string
          minLength: 6
        zonePreferee:
          type: string
          description: Optionnel - uniquement pour les livreurs
...
    CommentaireRequest:
      type: object
      properties:
        commentaire:
          type: string
    LivrerRequest:
      type: object
      required:
        - cashCollecte
      properties:
        cashCollecte:
          type: number
          format: decimal
          description: Montant réellement encaissé auprès du client
        commentaire:
          type: string
    EchecRequest:
      type: object
      required:
        - motif
      properties:
        motif:
          $ref: '#/components/schemas/MotifEchec'
        commentaire:
          type: string

    # ============================================
    # ERROR SCHEMA
    # ============================================
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string

    # ============================================
    # SCHEMAS - Admin Zones
    # ============================================

    # DTOs pour Admin Zones
...